% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{A numeric vector, degrees to be converted}
}
\value{
The radians converted from degrees
}
\description{
Convert degrees to radians
}
\examples{
rad <- deg2rad(180)
rads <- deg2rad(seq(0, 360, 90))
}

