% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.R
\name{getarc}
\alias{getarc}
\title{Calculate length of arc of a circle}
\usage{
getarc(rad, radius)
}
\arguments{
\item{rad}{degree in radian}

\item{radius}{radius of the circle}
}
\value{
Length of arc of a circle.
}
\description{
Calculate length of arc of a circle
}
\examples{
arc <- getarc(pi, 1)
arcs <- getarc(c(0, pi, 2*pi), 1:3)
}

