% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.R
\name{getarea}
\alias{getarea}
\title{Calculate area from different variables}
\usage{
getarea(variable = c("radius", "diameter", "circumference")[1], value)
}
\arguments{
\item{variable}{A character with name of variable supplied}

\item{value}{A numeric vector corresponds to the variable name}
}
\value{
Circular area calculated from supplied variable
}
\description{
Calculate area from different variables
}
\examples{
area <- getarea("radius", 1)
areas <- getarea("diameter", 1:10)
}

