% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.R
\name{getcircumference}
\alias{getcircumference}
\title{Calculate circumference from different variables}
\usage{
getcircumference(variable = c("radius", "diameter", "area")[1], value)
}
\arguments{
\item{variable}{A character with name of variable supplied}

\item{value}{A numeric vector corresponds to the variable name}
}
\value{
Circumference calculated from supplied variable
}
\description{
Calculate circumference from different variables
}
\examples{
circum <- getcircumference("area", pi)
circums <- getcircumference("radius", 1:10)
}

