% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.R
\name{getdiameter}
\alias{getdiameter}
\title{Calculate diameter from different variables}
\usage{
getdiameter(variable = c("radius", "diameter", "circumference")[1], value)
}
\arguments{
\item{variable}{A character with name of variable supplied}

\item{value}{A numeric vector corresponds to the variable name}
}
\value{
Diameter calculated from supplied variable
}
\description{
Calculate diameter from different variables
}
\examples{
diameter <- getdiameter("area", pi)
diameters <- getdiameter("radius", 1:10)
}

