% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.R
\name{getradius}
\alias{getradius}
\title{Calculate radius from different variables}
\usage{
getradius(variable = c("radius", "diameter", "circumference")[1], value)
}
\arguments{
\item{variable}{A character with name of variable supplied}

\item{value}{A numeric vector corresponds to the variable name}
}
\value{
Radius calculated from supplied variable
}
\description{
Calculate radius from different variables
}
\examples{
radius <- getradius("area", pi)
radii <- getradius("circumference", 1:10)
}

