% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.R
\name{getsector}
\alias{getsector}
\title{Calculate area of sector of a circle}
\usage{
getsector(rad, radius)
}
\arguments{
\item{rad}{degree in radian}

\item{radius}{radius of the circle}
}
\value{
Area of sector of a circle
}
\description{
Calculate area of sector of a circle
}
\examples{
sct <- getsector(pi, 1)
scts <- getsector(c(0, pi, 2*pi), 1:3)
}

