% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.R
\name{getsegment}
\alias{getsegment}
\title{Calculate area of segment of a circle}
\usage{
getsegment(rad, radius)
}
\arguments{
\item{rad}{degree in radian}

\item{radius}{radius of the circle}
}
\value{
Area of segment of a circle
}
\description{
Calculate area of segment of a circle
}
\examples{
seg <- getsegment(pi, 1)
segs <- getsegment(c(0, pi, 2*pi), 1:3)
}

