% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/triangle.R
\name{isline}
\alias{isline}
\title{Determine if multiple points are on one line}
\usage{
isline(x, y)
}
\arguments{
\item{x}{x coordinates of points}

\item{y}{y coordinates of points}
}
\value{
logical. TRUE indicates on one line
}
\description{
Determine if multiple points are on one line
}
\examples{
isline(c(0,0), c(1,1)) # TRUE
isline(1:3, rnorm(3))  # probably FALSE
}

