% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/triangle.R
\name{isobtuse}
\alias{isobtuse}
\title{Determine if a triangle is obtuse}
\usage{
isobtuse(x, y)
}
\arguments{
\item{x}{x coordinates of vertices of a triangle}

\item{y}{y coordinates of vertices of a triangle}
}
\value{
logical. TRUE indicates the triangle is obtuse
}
\description{
Determine if a triangle is obtuse
}
\examples{
isobtuse(1:3, c(2, 8, 9)) # TRUE
isobtuse(c(0,0,3), c(0,4,0))  # FALSE
}

