% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/triangle.R
\name{isright}
\alias{isright}
\title{Determine if a triangle is right}
\usage{
isright(x, y)
}
\arguments{
\item{x}{x coordinates of vertices of a triangle}

\item{y}{y coordinates of vertices of a triangle}
}
\value{
logical. TRUE indicates the triangle is right
}
\description{
Determine if a triangle is right
}
\examples{
isright(1:3, c(2, 8, 9)) # FALSE
isright(c(0,0,3), c(0,4,0))  # TRUE
}

