% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/draw.R
\name{plotcircle}
\alias{plotcircle}
\title{Plot circle with different methods}
\usage{
plotcircle(x = 0, y = 0, r = 1, npoints = 100, method = "plot",
  asp = 1, xlab = "x", ylab = "y", col = "black", ...)
}
\arguments{
\item{x,y}{coordinates of circle center}

\item{r}{radius of circle}

\item{npoints}{number of points}

\item{method}{plotting method}

\item{asp}{aspect default is 1}

\item{xlab,ylab}{x and y labels}

\item{col}{color of circle}
}
\value{
invisible NULL
}
\description{
Plot circle with different methods
}
\examples{
plotcircle(0,0,1)
}

