% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit-helper.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert radians to degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{A numeric vector, radians to be converted}
}
\value{
The degrees converted from radians
}
\description{
Convert radians to degrees
}
\examples{
deg <- rad2deg(pi)
degs <- rad2deg(seq(0, 2*pi, pi/2))
}

